move data_access into pipeline since the separation was pointless
